Use parent attribute to find inheritance chain
Fixes #13

While not documented, loggers have a `parent` attribute that determines
the inheritance chain. In this inheritance chain there are no
placeholder loggers, but the parent may be re-assigned by manually
assigning a different logger to the `parent` attribute.

To better reflect how the module actually operates, we need to be
following that `parent` attribute, rather than parsing the names. This
means that there will be no placeholders in the output, because the
`parent` never points to a placeholder, but rather skips over it
entirely.